+++ pypinyin/style/_utils.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from pypinyin.standard import convert_finals
 from pypinyin.style._constants import (
-    _INITIALS, _INITIALS_NOT_STRICT,
+    _INITIALS, _INITIALS_NOT_STRICT, _FINALS,
     RE_PHONETIC_SYMBOL, PHONETIC_SYMBOL_DICT,
     PHONETIC_SYMBOL_DICT_KEY_LENGTH_NOT_ONE,
     RE_NUMBER
@@ -33,7 +33,7 @@ def get_initials(pinyin, strict):
 def get_finals(pinyin, strict):
     """获取单个拼音中的韵母.
 
-    :param pinyin: 单个拼音
+    :param pinyin: 单个拼音，无声调拼音
     :type pinyin: unicode
     :param strict: 是否严格遵照《汉语拼音方案》来处理声母和韵母
     :return: 韵母
@@ -43,11 +43,24 @@ def get_finals(pinyin, strict):
         pinyin = convert_finals(pinyin)
 
     initials = get_initials(pinyin, strict=strict) or ''
-    # 没有声母，整个都是韵母
-    if not initials:
-        return pinyin
+
     # 按声母分割，剩下的就是韵母
-    return ''.join(pinyin.split(initials, 1))
+    finals = pinyin[len(initials):]
+
+    # 处理既没有声母也没有韵母的情况
+    if strict and finals not in _FINALS:
+        # 处理 y, w 导致误判的问题，比如 yo
+        initials = get_initials(pinyin, strict=False)
+        finals = pinyin[len(initials):]
+        if finals in _FINALS:
+            return finals
+        return ''
+
+    # ń, ḿ
+    if not finals and not strict:
+        return pinyin
+
+    return finals
 
 
 def replace_symbol_to_number(pinyin):
@@ -74,7 +87,7 @@ def replace_symbol_to_no_symbol(pinyin):
 def has_finals(pinyin):
     """判断是否有韵母"""
     # 鼻音: 'm̄', 'ḿ', 'm̀', 'ń', 'ň', 'ǹ ' 没有韵母
-    for symbol in ['m̄', 'ḿ', 'm̀', 'ń', 'ň', 'ǹ', 'ê̄', 'ế', 'ê̌', 'ề']:
+    for symbol in ['m̄', 'ḿ', 'm̀', 'ń', 'ň', 'ǹ']:
         if symbol in pinyin:
             return False